+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `laravel-mssql-changes` will be documented in this file.
 
  - Sort ShowChanges output by Change Version (without grouping by Table)
  - Create a test database for use with tests
+ - Added DisableTableChangeTracking action and command
 
 ## 1.0.0 - 2022-10-10